Move location-link-utils.ts to /common
