+++ extensions/ql-vscode/src/common/location-link-utils.ts
