+++ extensions/ql-vscode/src/variant-analysis/markdown-generation.ts
@@ -1,6 +1,6 @@
 import { CellValue } from "../common/bqrs-cli-types";
 import { tryGetRemoteLocation } from "../common/bqrs-utils";
-import { createRemoteFileRef } from "../pure/location-link-utils";
+import { createRemoteFileRef } from "../common/location-link-utils";
 import { parseHighlightedLine, shouldHighlightLine } from "../pure/sarif-utils";
 import { convertNonPrintableChars } from "../pure/text-utils";
 import type {
